feat: JWT 토큰을 통한 회원 탈퇴 기능 추가
- 회원 탈퇴 시 JWT 토큰을 통한 사용자 인증 로직 추가
- 탈퇴 요청 시 토큰 검증 및 유효성 검사 추가
- 토큰에서 사용자 정보를 추출해 요청한 사용자와 일치 여부 확인
- 본인만 회원 탈퇴를 할 수 있도록 제한
- 테스트 코드 수정 및 JWT 관련 테스트 추가

Refs: #18

+++ src/main/java/com/sparta/newsfeedproject/domain/member/service/MemberService.java
@@ -1,7 +1,10 @@
 package com.sparta.newsfeedproject.domain.member.service;
 
 import com.sparta.newsfeedproject.domain.member.entity.Member;
+import com.sparta.newsfeedproject.domain.member.jwt.JwtUtil;
 import com.sparta.newsfeedproject.domain.member.repository.MemberRepository;
+import io.jsonwebtoken.Claims;
+import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
@@ -10,13 +13,32 @@ import org.springframework.stereotype.Service;
 public class MemberService {
 
     private final MemberRepository memberRepository;
-    // 회원탈퇴
-    public void deleteMember(Long memberId) {
-        // 사용자 존재 여부 확인
-        Member member = memberRepository.findById(memberId).orElseThrow(() -> new IllegalArgumentException("사용자를 찾을 수 없습니다."));
+    private final JwtUtil jwtUtil;
 
-        // 사용자 삭제
-        memberRepository.delete(member);
+
+    // 회원탈퇴
+    public void deleteMember(Long memberId, HttpServletRequest request) {
+        // 요청 헤더에서 Jwt 토큰 추출
+        String token = jwtUtil.resolveToken(request);
+
+        // 토큰 유효성 겁사 및 사용자 정보 추출
+        if (token != null && jwtUtil.validateToken(token)) {
+            Claims claims = jwtUtil.getUserInfoFromToken(token);  // JWT에서 사용자 정보 추출
+            String tokenUsername = claims.getSubject();  // 토큰에서 가져온 사용자 이름
+
+            // 사용자 존재 여부 확인
+            Member member = memberRepository.findById(memberId)
+                    .orElseThrow(() -> new IllegalArgumentException("사용자를 찾을 수 없습니다."));
+
+            // 토큰의 사용자 정보와 요청한 memberId의 사용자 정보가 일치하는지 확인
+            if (!member.getName().equals(tokenUsername)) {
+                throw new IllegalArgumentException("본인만 탈퇴할 수 있습니다.");
+            }
+            // 사용자 삭제
+            memberRepository.delete(member);
+        } else {
+            throw new IllegalArgumentException("유효하지 않은 토큰입니다.");
+        }
     }
 }
